requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
facenet-pytorch==2.5.2
idna==3.3
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
requests==2.27.1
sqlparse==0.4.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9